% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMRCE.R
\name{RMRCE}
\alias{RMRCE}
\title{RMRCE}
\usage{
RMRCE(X, Y, lambda, alpha = 5)
}
\arguments{
\item{X}{A numeric matrix of explanatory variables.}

\item{Y}{A numeric vector of response.}

\item{lambda}{Tuning parameter lambda that can be tuned using cross validation. See RMRCE_cv.}

\item{alpha}{Tuning parameter alpha that can be tuned using cross validation. See RMRCE_cv. Default value is 5.}
}
\value{
A numeric vector of regression coefficients.
}
\description{
Regularized Maximum Rank Correlation Estimator
}
\details{
This function fits RMRCE (Regularized Maximum Rank Correlation Estimator). It is recommended that the tuning parameters be tuned using cross validation (see RMRCE_cv). RMRCE may need some time to finish with large datasets.
}
\examples{
Y <- rnorm(100)
X <- cbind(Y+rnorm(100,sd=0.1),-0.5*Y+rnorm(100,sd=0.1),rnorm(100,sd=0.1))
fit <- RMRCE(X,Y,0.01,5)
}
\author{
Fang Han, Hongkai Ji, Zhicheng Ji, Honglang Wang <zji4@jhu.edu>
}
