% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMRCE_cv.R
\name{RMRCE_cv}
\alias{RMRCE_cv}
\title{RMRCE_cv}
\usage{
RMRCE_cv(X, Y, lambdavec, alphavec)
}
\arguments{
\item{X}{A numeric matrix of explanatory variables.}

\item{Y}{A numeric vector of response.}

\item{lambda}{A numeric vector of tuning parameter lambda that the testing error is to be computed.}

\item{alpha}{A numeric vector of tuning parameter alpha that the testing error is to be computed.}
}
\value{
A data.frame which has three columns: lambda, alpha and corresponding testing error. Lambda and alpha that yields the smallest testing error are selected as the optimal tuning parameters.
}
\description{
Cross validation for choosing RMRCE tuning parameters
}
\details{
This function performs a five-fold cross validation to choose the optimal tuning parameter lambda and alpha which yields the smallest testing error.
}
\examples{
Y <- rnorm(100)
X <- cbind(Y+rnorm(100,sd=0.1),-0.5*Y+rnorm(100,sd=0.1),rnorm(100,sd=0.1))
cv <- RMRCE_cv(X,Y,c(0.001,0.01),c(1,5))
optlambda <- cv[which.min(cv[,3]),1]
optalpha <- cv[which.min(cv[,3]),2]
fit <- RMRCE(X,Y,optlambda,optalpha)
}
\author{
Fang Han, Hongkai Ji, Zhicheng Ji, Honglang Wang <zji4@jhu.edu>
}
